CHAD-17068: zigbee-humidity-sensor lazy load sub drivers
