+++ drivers/SmartThings/zigbee-humidity-sensor/src/centralite-sensor/fingerprints.lua
@@ -1,0 +1,10 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local CENTRALITE_SENSOR_FINGERPRINTS = {
+  { mfr = "CentraLite", model = "3310-S" },
+  { mfr = "CentraLite", model = "3310-G" },
+  { mfr = "CentraLite", model = "3310" }
+}
+
+return CENTRALITE_SENSOR_FINGERPRINTS
