+++ drivers/SmartThings/zigbee-humidity-sensor/src/centralite-sensor/init.lua
@@ -1,16 +1,6 @@
--- Copyright 2022 SmartThings
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
+-- Copyright 2022 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
 
 local capabilities = require "st.capabilities"
 local clusters = require "st.zigbee.zcl.clusters"
@@ -23,20 +13,7 @@ local configurationMap = require "configurations"
 local HUMIDITY_CLUSTER_ID = 0xFC45
 local HUMIDITY_MEASURE_ATTR_ID = 0x0000
 
-local CENTRALITE_SENSOR_FINGERPRINTS = {
-  { mfr = "CentraLite", model = "3310-S" },
-  { mfr = "CentraLite", model = "3310-G" },
-  { mfr = "CentraLite", model = "3310" }
-}
 
-local function can_handle_centralite_sensor(opts, driver, device)
-  for _, fingerprint in ipairs(CENTRALITE_SENSOR_FINGERPRINTS) do
-    if device:get_manufacturer() == fingerprint.mfr and device:get_model() == fingerprint.model then
-      return true
-    end
-  end
-  return false
-end
 
 local function device_init(driver, device)
   device:remove_configured_attribute(clusters.RelativeHumidity.ID, clusters.RelativeHumidity.attributes.MeasuredValue.ID)
@@ -89,7 +66,7 @@ local centralite_sensor = {
       }
     }
   },
-  can_handle = can_handle_centralite_sensor
+  can_handle = require("centralite-sensor.can_handle"),
 }
 
 return centralite_sensor
