+++ drivers/SmartThings/zigbee-humidity-sensor/src/frient-sensor/air-quality/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local FRIENT_AIR_QUALITY_SENSOR_FINGERPRINTS = {
+  { mfr = "frient A/S", model = "AQSZB-110", subdriver = "airquality" }
+}
+
+return FRIENT_AIR_QUALITY_SENSOR_FINGERPRINTS
