+++ drivers/SmartThings/zigbee-humidity-sensor/src/frient-sensor/fingerprints.lua
@@ -1,0 +1,10 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local FRIENT_TEMP_HUMUDITY_SENSOR_FINGERPRINTS = {
+  { mfr = "frient A/S", model = "HMSZB-110" },
+  { mfr = "frient A/S", model = "HMSZB-120" },
+  { mfr = "frient A/S", model = "AQSZB-110" }
+}
+
+return FRIENT_TEMP_HUMUDITY_SENSOR_FINGERPRINTS
