+++ drivers/SmartThings/zigbee-humidity-sensor/src/frient-sensor/init.lua
@@ -1,39 +1,13 @@
--- Copyright 2025 SmartThings
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
 
-local battery_defaults = require "st.zigbee.defaults.battery_defaults"
-local configurationMap = require "configurations"
 local zcl_clusters = require "st.zigbee.zcl.clusters"
 local HumidityMeasurement = zcl_clusters.RelativeHumidity
 local TemperatureMeasurement = zcl_clusters.TemperatureMeasurement
 
-local FRIENT_TEMP_HUMUDITY_SENSOR_FINGERPRINTS = {
-  { mfr = "frient A/S", model = "HMSZB-110" },
-  { mfr = "frient A/S", model = "HMSZB-120" },
-  { mfr = "frient A/S", model = "AQSZB-110" }
-}
-
-local function can_handle_frient_sensor(opts, driver, device)
-  for _, fingerprint in ipairs(FRIENT_TEMP_HUMUDITY_SENSOR_FINGERPRINTS) do
-    if device:get_manufacturer() == fingerprint.mfr and device:get_model() == fingerprint.model then
-      return true
-    end
-  end
-  return false
-end
-
 local function device_init(driver, device)
+  local configurationMap = require "configurations"
+  local battery_defaults = require "st.zigbee.defaults.battery_defaults"
   battery_defaults.build_linear_voltage_init(2.3,3.0)(driver, device)
   local configuration = configurationMap.get_device_configuration(device)
   if configuration ~= nil then
@@ -74,10 +48,8 @@ local frient_sensor = {
     doConfigure = do_configure,
     infoChanged = info_changed
   },
-  sub_drivers = {
-    require("frient-sensor/air-quality")
-  },
-  can_handle = can_handle_frient_sensor
+  sub_drivers = require("frient-sensor.sub_drivers"),
+  can_handle = require("frient-sensor.can_handle"),
 }
 
 return frient_sensor
