+++ drivers/SmartThings/zigbee-humidity-sensor/src/frient-sensor/sub_drivers.lua
@@ -1,0 +1,9 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local lazy_load_if_possible = require "lazy_load_subdriver"
+local sub_drivers = {
+    lazy_load_if_possible("frient-sensor.air-quality")
+}
+
+return sub_drivers
