+++ drivers/SmartThings/zigbee-humidity-sensor/src/heiman-sensor/fingerprints.lua
@@ -1,0 +1,11 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local HEIMAN_TEMP_HUMUDITY_SENSOR_FINGERPRINTS = {
+  { mfr = "Heiman", model = "b467083cfc864f5e826459e5d8ea6079" },
+  { mfr = "HEIMAN", model = "888a434f3cfc47f29ec4a3a03e9fc442" },
+  { mfr = "HEIMAN", model = "HT-EM" },
+  { mfr = "HEIMAN", model = "HT-EF-3.0" }
+}
+
+return HEIMAN_TEMP_HUMUDITY_SENSOR_FINGERPRINTS
