+++ drivers/SmartThings/zigbee-humidity-sensor/src/init.lua
@@ -1,16 +1,6 @@
--- Copyright 2022 SmartThings
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
+-- Copyright 2022 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
 
 local capabilities = require "st.capabilities"
 local ZigbeeDriver = require "st.zigbee"
@@ -78,14 +68,7 @@ local zigbee_humidity_driver = {
     init = device_init,
     added = added_handler,
   },
-  sub_drivers = {
-    require("aqara"),
-    require("plant-link"),
-    require("plaid-systems"),
-    require("centralite-sensor"),
-    require("heiman-sensor"),
-    require("frient-sensor")
-  },
+  sub_drivers = require("sub_drivers"),
   health_check = false,
 }
