+++ drivers/SmartThings/zigbee-humidity-sensor/src/plaid-systems/fingerprints.lua
@@ -1,0 +1,10 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local ZIGBEE_HUMIDITY_SENSOR_FINGERPRINTS = {
+   { mfr = "PLAID SYSTEMS", model = "PS-SPRZMS-01" },
+   { mfr = "PLAID SYSTEMS", model = "PS-SPRZMS-SLP1" },
+   { mfr = "PLAID SYSTEMS", model = "PS-SPRZMS-SLP3" },
+}
+
+return ZIGBEE_HUMIDITY_SENSOR_FINGERPRINTS
