+++ drivers/SmartThings/zigbee-humidity-sensor/src/plaid-systems/init.lua
@@ -1,16 +1,6 @@
--- Copyright 2022 SmartThings
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
+-- Copyright 2022 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
 
 local capabilities = require "st.capabilities"
 local device_management = require "st.zigbee.device_management"
@@ -21,21 +11,7 @@ local PowerConfiguration = zcl_clusters.PowerConfiguration
 local RelativeHumidity = zcl_clusters.RelativeHumidity
 local TemperatureMeasurement = zcl_clusters.TemperatureMeasurement
 
-local ZIGBEE_HUMIDITY_SENSOR_FINGERPRINTS = {
-   { mfr = "PLAID SYSTEMS", model = "PS-SPRZMS-01" },
-   { mfr = "PLAID SYSTEMS", model = "PS-SPRZMS-SLP1" },
-   { mfr = "PLAID SYSTEMS", model = "PS-SPRZMS-SLP3" },
-}
 
-local is_zigbee_plaid_systems_humidity_sensor = function(opts, driver, device)
-  for _, fingerprint in ipairs(ZIGBEE_HUMIDITY_SENSOR_FINGERPRINTS) do
-    if device:get_manufacturer() == fingerprint.mfr and device:get_model() == fingerprint.model then
-      return true
-    end
-  end
-
-  return false
-end
 
 local battery_mains_voltage_attr_handler = function(driver, device, value, zb_rx)
   local min = 2500
@@ -87,7 +63,7 @@ local plaid_systems_humdity_sensor = {
       }
     }
   },
-  can_handle = is_zigbee_plaid_systems_humidity_sensor
+  can_handle = require("plaid-systems.can_handle"),
 }
 
 return plaid_systems_humdity_sensor
